# backend/src/database/models/Branch.js
import { DataTypes, Model } from 'sequelize';
import sequelize from '../Sequelize.js';

class Branch extends Model {}

Branch.init({
  // id is auto-generated by Sequelize
  name: {
    type: DataTypes.STRING,
    allowNull: false,
  },
  workingTerms: {
    type: DataTypes.JSON, // Stores the array of Term objects
    allowNull: false,
    defaultValue: [],
  },
  // projectId is added via associations
}, {
  sequelize,
  modelName: 'Branch',
  timestamps: false,
  indexes: [
    {
      unique: true,
      fields: ['name', 'projectId']
    }
  ]
});

export default Branch;