# openapi.json
{
    "openapi":"3.0.0","info":{"title":"Localization Manager Pro API","description":"API for managing localization projects, terms, translations, and teams. Based on the mock ApiClient.ts.","version":"1.0.0"},"servers":[{"url":"/api/v1","description":"Mock API Server"}],"tags":[{"name":"Authentication"},{"name":"Users"},{"name":"Projects"},{"name":"Terms"},{"name":"Translations"},{"name":"Team"},{"name":"Branches"},{"name":"Commits","description":"Operations related to commit history"}],"paths":{"/login":{"post":{"summary":"Log in a user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","pass"],"properties":{"email":{"type":"string","format":"email"},"pass":{"type":"string","format":"password"}}}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized / Invalid credentials"}}}},"/users":{"get":{"summary":"Get all users","tags":["Users"],"responses":{"200":{"description":"A list of all users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/users/me/profile":{"put":{"summary":"Update current user's profile","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}}}}},"responses":{"200":{"description":"Profile updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized"}}}},"/users/me/password":{"put":{"summary":"Change current user's password","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["currentPassword","newPassword"],"properties":{"currentPassword":{"type":"string","format":"password"},"newPassword":{"type":"string","format":"password"}}}}}},"responses":{"200":{"description":"Password updated successfully"},"400":{"description":"Invalid request (e.g., current password incorrect)"},"401":{"description":"Unauthorized"}}}},"/projects":{"get":{"summary":"Get all projects","tags":["Projects"],"responses":{"200":{"description":"A list of all projects","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}}},"post":{"summary":"Create a new project","tags":["Projects"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name","userId"],"properties":{"name":{"type":"string"},"userId":{"type":"string"}}}}}},"responses":{"201":{"description":"Project created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}}}}},"/projects/{projectId}/languages":{"put":{"summary":"Update project languages","tags":["Projects"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Language"}}}}},"responses":{"200":{"description":"Languages updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"404":{"description":"Project not found"}}}},"/projects/{projectId}/defaultLanguage":{"put":{"summary":"Set the default language for a project","tags":["Projects"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["langCode"],"properties":{"langCode":{"type":"string"}}}}}},"responses":{"200":{"description":"Default language updated"},"404":{"description":"Project not found"}}}},"/projects/{projectId}/branches/{branchName}/terms":{"post":{"summary":"Add a new term to a branch","tags":["Terms"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["termText"],"properties":{"text":{"type":"string"}}}}}},"responses":{"201":{"description":"Term created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Term"}}}},"404":{"description":"Project or branch not found"}}}},"/projects/{projectId}/branches/{branchName}/terms/{termId}":{"put":{"summary":"Update a term's key text","tags":["Terms"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}},{"name":"termId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["text"],"properties":{"text":{"type":"string"}}}}}},"responses":{"200":{"description":"Term updated"},"404":{"description":"Project, branch, or term not found"}}},"delete":{"summary":"Delete a term from a branch","tags":["Terms"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}},{"name":"termId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Term deleted successfully"},"404":{"description":"Project, branch, or term not found"}}}},"/projects/{projectId}/branches/{branchName}/terms/{termId}/context":{"put":{"summary":"Update a term's context","tags":["Terms"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}},{"name":"termId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["context"],"properties":{"context":{"type":"string"}}}}}},"responses":{"200":{"description":"Context updated"},"404":{"description":"Project, branch, or term not found"}}}},"/projects/{projectId}/branches/{branchName}/terms/{termId}/translations/{langCode}":{"put":{"summary":"Update a translation for a term","tags":["Translations"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}},{"name":"termId","in":"path","required":true,"schema":{"type":"string"}},{"name":"langCode","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["value"],"properties":{"value":{"type":"string"}}}}}},"responses":{"200":{"description":"Translation updated"},"404":{"description":"Project, branch, or term not found"}}}},"/projects/{projectId}/team":{"post":{"summary":"Add a member to a project","tags":["Team"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","role"],"properties":{"email":{"type":"string","format":"email"},"role":{"$ref":"#/components/schemas/UserRole"},"languages":{"type":"array","items":{"type":"string"},"description":"Optional initial languages to assign to the new member."}}}}}},"responses":{"200":{"description":"Response indicating success or failure of the operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddMemberResult"}}}},"404":{"description":"Project or user not found"},"409":{"description":"User is already a member"}}}},"/projects/{projectId}/team/{userId}":{"delete":{"summary":"Remove a member from a project","tags":["Team"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Member removed"},"404":{"description":"Project or member not found"}}}},"/projects/{projectId}/team/{userId}/role":{"put":{"summary":"Update a team member's role","tags":["Team"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["role"],"properties":{"role":{"$ref":"#/components/schemas/UserRole"}}}}}},"responses":{"200":{"description":"Role updated"},"404":{"description":"Project or member not found"}}}},"/projects/{projectId}/team/{userId}/languages":{"put":{"summary":"Update languages assigned to a team member","tags":["Team"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"A list of language codes","required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":"Assigned languages updated"},"404":{"description":"Project or member not found"}}}},"/projects/{projectId}/branches":{"post":{"summary":"Create a new branch from another branch","tags":["Branches"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["newBranchName","sourceBranchName"],"properties":{"newBranchName":{"type":"string"},"sourceBranchName":{"type":"string"}}}}}},"responses":{"201":{"description":"Branch created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Branch"}}}},"404":{"description":"Project or source branch not found"},"409":{"description":"Branch name already exists"}}}},"/projects/{projectId}/branches/from-commit":{"post":{"summary":"Create a new branch from a specific commit","tags":["Branches"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["commitId","newBranchName"],"properties":{"commitId":{"type":"string"},"newBranchName":{"type":"string"}}}}}},"responses":{"201":{"description":"Branch created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Branch"}}}},"404":{"description":"Project or commit not found"},"409":{"description":"Branch name already exists"}}}},"/projects/{projectId}/branches/{branchName}":{"delete":{"summary":"Delete a branch","tags":["Branches"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Branch deleted"},"400":{"description":"Cannot delete the main branch"},"404":{"description":"Project or branch not found"}}}},"/projects/{projectId}/currentBranch":{"put":{"summary":"Switch the current branch for a project","tags":["Branches"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["branchName"],"properties":{"branchName":{"type":"string"}}}}}},"responses":{"200":{"description":"Branch switched"},"404":{"description":"Project or branch not found"}}}},"/projects/{projectId}/branches/merge":{"post":{"summary":"Merge one branch into another","tags":["Branches"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["sourceBranchName","targetBranchName"],"properties":{"sourceBranchName":{"type":"string"},"targetBranchName":{"type":"string"}}}}}},"responses":{"200":{"description":"Merge successful"},"404":{"description":"Project or one of the branches not found"}}}},"/projects/{projectId}/branches/{branchName}/commits":{"post":{"summary":"Create a new commit","tags":["Commits"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["message","authorId"],"properties":{"message":{"type":"string"},"authorId":{"type":"string"}}}}}},"responses":{"201":{"description":"Commit created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Commit"}}}},"404":{"description":"Project or branch not found"}}}},"/projects/{projectId}/branches/{branchName}/commits/latest":{"delete":{"summary":"Delete the latest commit of a branch","tags":["Commits"],"parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"branchName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Commit deleted successfully"},"400":{"description":"Cannot delete the initial commit"},"404":{"description":"Project or branch not found"}}}}},"components":{"schemas":{"Language":{"type":"object","properties":{"code":{"type":"string","example":"en"},"name":{"type":"string","example":"English"}}},"Term":{"type":"object","properties":{"id":{"type":"string"},"text":{"type":"string","description":"The term key."},"context":{"type":"string","description":"Context for translators."},"translations":{"type":"object","additionalProperties":{"type":"string"},"example":{"en":"Submit","es":"Enviar"}}}},"Commit":{"type":"object","properties":{"id":{"type":"string"},"message":{"type":"string"},"authorId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"terms":{"type":"array","items":{"$ref":"#/components/schemas/Term"}}}},"Branch":{"type":"object","properties":{"name":{"type":"string"},"commits":{"type":"array","items":{"$ref":"#/components/schemas/Commit"}},"workingTerms":{"type":"array","items":{"$ref":"#/components/schemas/Term"}}}},"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string","format":"email"},"avatarInitials":{"type":"string"}}},"UserRole":{"type":"string","enum":["translator","editor","admin"]},"Project":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"defaultLanguageCode":{"type":"string"},"languages":{"type":"array","items":{"$ref":"#/components/schemas/Language"}},"branches":{"type":"array","items":{"$ref":"#/components/schemas/Branch"}},"currentBranchName":{"type":"string"},"team":{"type":"object","description":"Maps user ID to role and assigned languages.","additionalProperties":{"type":"object","properties":{"role":{"$ref":"#/components/schemas/UserRole"},"languages":{"type":"array","items":{"type":"string"}}}}}}},"AddMemberResult":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"success":{"type":"boolean"},"message":{"type":"string"},"code":{"type":"string","enum":["user_exists","user_not_found","project_not_found"]}}}}}}